spike(html): remove dependence on XMLDocument
Move methods inherited from XMLDocument into HTMLDocument and subclass
the parent (Document) instead. These inherited methods can then be
refactored and largely removed by later steps.

XMLDocument and its module are dead code now, remove it and its tests.
